KRACOEUS-6956 : move MailBy and MailType from propdev mailing to common; Keep KeyPersonInfo with propdev print for now
