+++ coeus-code/src/main/java/org/kuali/coeus/common/proposal/framework/mail/MailBy.java
@@ -1,29 +1,29 @@
 /*
  * Copyright 2005-2014 The Kuali Foundation
- * 
- * Licensed under the Educational Community License, Version 2.0 (the "License");
+ *
+ * Licensed under the Educational Community License, Version 2.0 (the License);
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.osedu.org/licenses/ECL-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.coeus.propdev.impl.mailing;
+package org.kuali.coeus.common.proposal.framework.mail;
 
 import org.kuali.coeus.sys.framework.model.KcPersistableBusinessObjectBase;
 
 /**
  * 
- * This class the bo class of mail type.
+ * This class is the bo class of mailby code.
  */
-public class MailType extends KcPersistableBusinessObjectBase {
+public class MailBy extends KcPersistableBusinessObjectBase {
 
-    private String mailType;
+    private String mailByCode;
 
     private String description;
 
@@ -35,11 +35,11 @@ public class MailType extends KcPersistableBusinessObjectBase {
         this.description = description;
     }
 
-    public String getMailType() {
-        return mailType;
+    public String getMailByCode() {
+        return mailByCode;
     }
 
-    public void setMailType(String mailType) {
-        this.mailType = mailType;
+    public void setMailByCode(String mailByCode) {
+        this.mailByCode = mailByCode;
     }
 }
