+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/print/AbstractResearchAndRelatedStream.java
@@ -35,7 +35,6 @@ import org.kuali.kra.proposaldevelopment.bo.DevelopmentProposal;
 import org.kuali.coeus.propdev.impl.person.ProposalPerson;
 import org.kuali.coeus.propdev.impl.location.ProposalSite;
 import org.kuali.kra.proposaldevelopment.printing.xmlstream.CompensationInfo;
-import org.kuali.kra.proposaldevelopment.printing.xmlstream.KeyPersonInfo;
 import org.kuali.kra.proposaldevelopment.specialreview.ProposalSpecialReview;
 import org.kuali.rice.krad.service.BusinessObjectService;
