+++ coeus-code/src/main/resources/org/kuali/kra/datadictionary/MailType.xml
@@ -2,7 +2,7 @@
 <beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:p="http://www.springframework.org/schema/p" xsi:schemaLocation="http://www.springframework.org/schema/beans         http://www.springframework.org/schema/beans/spring-beans.xsd">
   <bean id="MailType" parent="MailType-parentBean" />
   <bean id="MailType-parentBean" abstract="true" parent="BusinessObjectEntry">
-    <property name="businessObjectClass" value="org.kuali.coeus.propdev.impl.mailing.MailType" />
+    <property name="businessObjectClass" value="org.kuali.coeus.common.proposal.framework.mail.MailType" />
 		<property name="objectLabel" value="Mail Type" />
     <property name="inquiryDefinition" >
       <ref bean="MailType-inquiryDefinition" />
@@ -74,7 +74,7 @@
 
   <bean id="MailType-InquiryView-parentBean" abstract="true" parent="Uif-InquiryView">
     <property name="headerText" value="Mail Type"/>
-    <property name="dataObjectClassName" value="org.kuali.coeus.propdev.impl.mailing.MailType"/>
+    <property name="dataObjectClassName" value="org.kuali.coeus.common.proposal.framework.mail.MailType"/>
     <property name="items">
       <list>
         <bean parent="Uif-Disclosure-GridSection">
@@ -118,7 +118,7 @@
   </bean>
 
   <bean id="MailType-LookupView-parentBean" abstract="true" parent="Uif-LookupView">
-    <property name="dataObjectClassName" value="org.kuali.coeus.propdev.impl.mailing.MailType"/>
+    <property name="dataObjectClassName" value="org.kuali.coeus.common.proposal.framework.mail.MailType"/>
     <property name="headerText" value="Mail Type Lookup"/>
     <property name="page.header.lowerGroup">
       <bean parent="Uif-HeaderLowerGroup">
