+++ coeus-code/src/main/resources/org/kuali/kra/datadictionary/docs/MailByMaintenanceDocument.xml
@@ -3,7 +3,7 @@
   <bean id="MailByMaintenanceDocument" parent="MailByMaintenanceDocument-parentBean"/>
 
   <bean id="MailByMaintenanceDocument-parentBean" abstract="true" parent="KcMaintenanceDocumentEntry">
-    <property name="businessObjectClass" value="org.kuali.coeus.propdev.impl.mailing.MailBy"/>
+    <property name="businessObjectClass" value="org.kuali.coeus.common.proposal.framework.mail.MailBy"/>
     <property name="maintainableClass" value="org.kuali.kra.maintenance.KraMaintainableImpl"/>
     <property name="maintainableSections">
       <list>
