+++ coeus-code/src/main/resources/repository.xml
@@ -222,7 +222,7 @@
     	
 	</class-descriptor>
 
-	<class-descriptor class="org.kuali.coeus.propdev.impl.mailing.MailBy" table="MAIL_BY">
+	<class-descriptor class="org.kuali.coeus.common.proposal.framework.mail.MailBy" table="MAIL_BY">
 		<field-descriptor name="mailByCode" column="MAIL_BY_CODE" jdbc-type="VARCHAR" primarykey="true" />
 		<field-descriptor name="description" column="DESCRIPTION" jdbc-type="VARCHAR" />
 		<field-descriptor name="updateTimestamp" column="UPDATE_TIMESTAMP" jdbc-type="TIMESTAMP" />
@@ -346,7 +346,7 @@
 		<field-descriptor name="updateUser" column="UPDATE_USER" jdbc-type="VARCHAR"/>
 		<field-descriptor name="versionNumber" column="VER_NBR"  jdbc-type="BIGINT" locking="false"/>
 	</class-descriptor>
-	<class-descriptor class="org.kuali.coeus.propdev.impl.mailing.MailType" table="MAIL_TYPE">
+	<class-descriptor class="org.kuali.coeus.common.proposal.framework.mail.MailType" table="MAIL_TYPE">
 		<field-descriptor name="mailType" column="MAIL_TYPE" jdbc-type="VARCHAR" primarykey="true" />
 		<field-descriptor name="description" column="DESCRIPTION" jdbc-type="VARCHAR" />
 		<field-descriptor name="updateTimestamp" column="UPDATE_TIMESTAMP" jdbc-type="TIMESTAMP" />
