+++ coeus-it/src/test/java/org/kuali/kra/bo/KraPersistableBusinessObjectBaseTest.java
@@ -19,7 +19,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.kuali.coeus.sys.framework.model.KcPersistableBusinessObjectBase;
-import org.kuali.coeus.propdev.impl.mailing.MailType;
+import org.kuali.coeus.common.proposal.framework.mail.MailType;
 import org.kuali.kra.test.infrastructure.KcIntegrationTestBase;
 import org.kuali.rice.krad.UserSession;
 import org.kuali.rice.krad.util.GlobalVariables;
