+++ coeus-webapp/src/main/webapp/WEB-INF/tags/proposaldevelopment/proposalDevelopmentDeliveryInfo.tag
@@ -23,7 +23,7 @@
 	<div class="tab-container" align="center">
     	<h3>
     		<span class="subhead-left">Delivery Info</span>
-    		<span class="subhead-right"><kul:help businessObjectClassName="org.kuali.coeus.propdev.impl.mailing.MailBy" altText="help"/></span>
+    		<span class="subhead-right"><kul:help businessObjectClassName="org.kuali.coeus.common.proposal.framework.mail.MailBy" altText="help"/></span>
         </h3>
         <table cellpadding=0 cellspacing=0 summary="">
              <tr>
